+++ src/main/java/com/sejong/projectservice/infrastructure/document/entity/DocumentEntity.java
@@ -84,7 +84,6 @@ public class DocumentEntity {
                 .content(content)
                 .createdAt(createdAt)
                 .updatedAt(updatedAt)
-                .projectId(projectEntity != null ? projectEntity.getId() : null)
                 .build();
     }
